refactor Vehicle Entity for more controle

+++ api/src/Service/VehicleService.php
@@ -16,7 +16,7 @@ class VehicleService
     {
         $this->configLoaderService = $configLoaderService;
     }
-    public function getVehicleWithCost(int $basePrice, String $type): Vehicle
+    public function getVehicleWithCost(float $basePrice, String $type): Vehicle
     {
         $vehicle = new Vehicle($basePrice, $type);
